+++ api/client/src/main/java/org/projectnessie/client/auth/oauth2/DeviceCodeFlow.java
@@ -130,12 +130,7 @@ class DeviceCodeFlow extends AbstractFlow {
   }
 
   private DeviceCodeResponse requestDeviceCode() {
-    DeviceCodeRequest body =
-        ImmutableDeviceCodeRequest.builder()
-            // don't include client id, it's in the basic auth header
-            .scope(config.getScope().orElse(null))
-            .build();
-    return invokeEndpoint(config.getResolvedDeviceAuthEndpoint(), body, DeviceCodeResponse.class);
+    return invokeDeviceAuthEndpoint();
   }
 
   private void checkPollInterval(Duration serverPollInterval) {
@@ -165,12 +160,8 @@ class DeviceCodeFlow extends AbstractFlow {
   private void pollForNewTokens(String deviceCode) {
     try {
       LOGGER.debug("Device Code Flow: polling for new tokens");
-      DeviceCodeTokensRequest request =
-          ImmutableDeviceCodeTokensRequest.builder()
-              .deviceCode(deviceCode)
-              .scope(config.getScope().orElse(null))
-              // don't include client id, it's in the basic auth header
-              .build();
+      DeviceCodeTokensRequest.Builder request =
+          ImmutableDeviceCodeTokensRequest.builder().deviceCode(deviceCode);
       Tokens tokens = invokeTokenEndpoint(request, DeviceCodeTokensResponse.class);
       LOGGER.debug("Device Code Flow: new tokens received");
       tokensFuture.complete(tokens);