+++ api/client/src/main/java/org/projectnessie/client/auth/oauth2/DeviceCodeRequest.java
@@ -41,4 +41,12 @@ interface DeviceCodeRequest {
   @Nullable
   @JsonProperty("scope")
   String getScope();
+
+  interface Builder {
+    Builder clientId(String clientId);
+
+    Builder scope(String scope);
+
+    DeviceCodeRequest build();
+  }
 }