+++ api/client/src/main/java/org/projectnessie/client/auth/oauth2/DeviceCodeTokensRequest.java
@@ -18,7 +18,6 @@ package org.projectnessie.client.auth.oauth2;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import javax.annotation.Nullable;
 import org.immutables.value.Value;
 
 /**
@@ -40,10 +39,10 @@ import org.immutables.value.Value;
 @Value.Immutable
 @JsonSerialize(as = ImmutableDeviceCodeTokensRequest.class)
 @JsonDeserialize(as = ImmutableDeviceCodeTokensRequest.class)
-interface DeviceCodeTokensRequest extends TokensRequestBase {
+interface DeviceCodeTokensRequest extends PublicTokensRequestBase {
 
   /** REQUIRED. Value MUST be set to "urn:ietf:params:oauth:grant-type:device_code" */
-  @Value.Default
+  @Value.Derived
   @JsonProperty("grant_type")
   @Override
   default GrantType getGrantType() {
@@ -57,11 +56,7 @@ interface DeviceCodeTokensRequest extends TokensRequestBase {
   @JsonProperty("device_code")
   String getDeviceCode();
 
-  /**
-   * The client identifier as described in Section 2.2 of [RFC6749]. REQUIRED if the client is not
-   * authenticating with the authorization server as described in Section 3.2.1. of [RFC6749].
-   */
-  @JsonProperty("client_id")
-  @Nullable
-  String getClientId();
+  interface Builder extends PublicTokensRequestBase.Builder<DeviceCodeTokensRequest> {
+    Builder deviceCode(String deviceCode);
+  }
 }